Minor fixes in smooth torque, and format ecu.cpp

+++ ECU_V2/lib/common/ecu.hpp
@@ -17,7 +17,7 @@ public:
     std::optional<CAN_message_t> emitMessage(uint32_t current_time_ms);
 
     /* smooth torque */
-    int16_t smoothTorque(uint16_t current_time, uint16_t torque);
+    int16_t smoothTorque(uint32_t current_time, int16_t torque);
 
     /* Uses PRINTF for all printing. */
     void printState();
